feat(users): единая кнопка "Права доступа" с табами роль/группы и индивидуальные права (#410)
В деталях пользователя было три кнопки (История, Права доступа, Роль и группы) -
не помещались. Объединил "Права доступа" и "Роль и группы" в одну кнопку, которая
открывает модалку с двумя вкладками: "Роль и группы" (роль/группы/бан) и
"Индивидуальные права" (дерево ключей).

Тело UserPermissionsModal вынес в RolesGroupsPanel, сам UserPermissionsModal стал
тонкой обёрткой (контракт для AdminUsers сохранён). Дерево прав вынес в
IndividualPermissionsPanel. Новая UserAccessModal - табовая оболочка по эталону
(Teleport, useOverlayClose, Escape, анимация закрытия, radius 30px). Уведомления
о сохранении и ошибках через useDeletionsStore.
